+++ cobspec/src/test/java/com/saralein/cobspec/controller/DefaultControllerTest.java
@@ -1,11 +1,11 @@
 package com.saralein.cobspec.controller;
 
+import com.saralein.server.exchange.Header;
 import com.saralein.server.request.Request;
-import com.saralein.server.response.Header;
 import com.saralein.server.response.Response;
-
-import static org.junit.Assert.*;
 import org.junit.Test;
+import static org.junit.Assert.assertArrayEquals;
+import static org.junit.Assert.assertEquals;
 
 public class DefaultControllerTest {
     @Test
