+++ server/src/test/java/com/saralein/server/middleware/StaticMiddlewareTest.java
@@ -1,21 +1,19 @@
 package com.saralein.server.middleware;
 
 import com.saralein.server.FileHelper;
+import com.saralein.server.exchange.Header;
 import com.saralein.server.mocks.MockCallable;
 import com.saralein.server.mocks.MockHandler;
 import com.saralein.server.request.Request;
-import com.saralein.server.response.Header;
 import com.saralein.server.response.Response;
 import org.junit.Before;
 import org.junit.Test;
 import org.junit.runner.RunWith;
 import org.junit.runners.Parameterized;
-
 import java.nio.file.Path;
 import java.nio.file.Paths;
 import java.util.Arrays;
 import java.util.Collection;
-
 import static org.junit.Assert.assertArrayEquals;
 import static org.junit.Assert.assertEquals;
