+++ server/src/test/java/com/saralein/server/response/ResponseTest.java
@@ -1,8 +1,10 @@
 package com.saralein.server.response;
 
+import com.saralein.server.exchange.Header;
 import org.junit.Before;
 import org.junit.Test;
-import static org.junit.Assert.*;
+import static org.junit.Assert.assertArrayEquals;
+import static org.junit.Assert.assertEquals;
 
 public class ResponseTest {
     private Response response;
